+++ power_manager/devices/KW47B42ZB7/fsl_pm_device.h
@@ -341,7 +341,7 @@ enum _pm_resc_name
  * @{
  */
 #define PM_WSID_PTC6_RISING_EDGE PM_ENCODE_WAKEUP_SOURCE_ID(0UL, 11UL, PORTC_EFT_IRQn, 1UL)
-#define PM_WSID_PTC6_FALLING_EDGE PM_ENCODE_WAKEUP_SOURCE_ID(0UL, 11UL, PORTC_EFT_IRQn, 2UL)
+#define PM_WSID_PTC6_FALLING_EDGE (uint32_t) PM_ENCODE_WAKEUP_SOURCE_ID(0UL, 11UL, (uint32_t) PORTC_EFT_IRQn, 2UL)
 #define PM_WSID_PTC6_ANY_EDGE PM_ENCODE_WAKEUP_SOURCE_ID(0UL, 11UL, PORTC_EFT_IRQn, 3UL)
 /*! @} */
 
@@ -522,7 +522,7 @@ enum _pm_resc_name
 /*!
  * @brief Enable GPIOD Low pins as wakeup source.
  */
-#define PM_WSID_GPIOD_LOW PM_ENCODE_WAKEUP_SOURCE_ID(1UL, 1UL, GPIOD_INT0_IRQn, 0UL)
+#define PM_WSID_GPIOD_LOW (uint32_t) PM_ENCODE_WAKEUP_SOURCE_ID(1UL, 1UL, (uint32_t) GPIOD_INT0_IRQn, 0UL)
 
 /*!
  * @brief Enable GPIOD High pins as wakeup source.
